✨ feat(damageNumber): TMP/伤害数字浮动

+++ client/Assets/Scripts/GameMain/Entity/EntityLogic/EntityPlayer.cs
@@ -52,7 +52,7 @@ namespace p1
             _entityDataPlayer.Anim.SetBool("isMoving", moveInput != Vector3.zero);
         }
         
-        public void GetDamage(int damage)
+        public void GetDamage(float damage)
         {
             if (_entityDataPlayer == null)
             {
